Removed support for memcache (without D)
