+++ framework/caching/MemCached.php
@@ -1,0 +1,285 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\caching;
+
+use Yii;
+use yii\base\InvalidConfigException;
+
+/**
+ * MemCached implements a cache application component based on [memcached](http://pecl.php.net/package/memcached) PECL
+ * extension.
+ *
+ * MemCached can be configured with a list of memcached servers by settings its [[servers]] property.
+ * By default, MemCached assumes there is a memcached server running on localhost at port 11211.
+ *
+ * See [[Cache]] for common cache operations that MemCached supports.
+ *
+ * Note, there is no security measure to protected data in memcached.
+ * All data in memcached can be accessed by any process running in the system.
+ *
+ * To use MemCached as the cache application component, configure the application as follows,
+ *
+ * ```php
+ * [
+ *     'components' => [
+ *         'cache' => [
+ *             'class' => \yii\caching\MemCached::class,
+ *             'servers' => [
+ *                 [
+ *                     'host' => 'server1',
+ *                     'port' => 11211,
+ *                     'weight' => 60,
+ *                 ],
+ *                 [
+ *                     'host' => 'server2',
+ *                     'port' => 11211,
+ *                     'weight' => 40,
+ *                 ],
+ *             ],
+ *         ],
+ *     ],
+ * ]
+ * ```
+ *
+ * In the above, two memcached servers are used: server1 and server2. You can configure more properties of
+ * each server, such as `persistent`, `weight`, `timeout`. Please see [[MemCacheServer]] for available options.
+ *
+ * For more details and usage information on Cache, see the [guide article on caching](guide:caching-overview).
+ *
+ * @property \Memcached $memcached The memcached object used by this cache component.
+ * This property is read-only.
+ * @property MemCachedServer[] $servers List of memcached server configurations. Note that the type of this
+ * property differs in getter and setter. See [[getServers()]] and [[setServers()]] for details.
+ *
+ * @author Qiang Xue <qiang.xue@gmail.com>
+ * @since 2.0
+ */
+class MemCached extends Cache
+{
+    /**
+     * @var string an ID that identifies a Memcached instance.
+     * By default the Memcached instances are destroyed at the end of the request. To create an instance that
+     * persists between requests, you may specify a unique ID for the instance. All instances created with the
+     * same ID will share the same connection.
+     * @see http://ca2.php.net/manual/en/memcached.construct.php
+     */
+    public $persistentId;
+    /**
+     * @var array options for Memcached.
+     * @see http://ca2.php.net/manual/en/memcached.setoptions.php
+     */
+    public $options;
+    /**
+     * @var string memcached sasl username.
+     * @see http://php.net/manual/en/memcached.setsaslauthdata.php
+     */
+    public $username;
+    /**
+     * @var string memcached sasl password.
+     * @see http://php.net/manual/en/memcached.setsaslauthdata.php
+     */
+    public $password;
+
+    /**
+     * @var \Memcached the Memcached instance
+     */
+    private $_cache;
+    /**
+     * @var array list of memcached server configurations
+     */
+    private $_servers = [];
+
+
+    /**
+     * Initializes this application component.
+     * It creates the memcached instance and adds memcached servers.
+     */
+    public function init()
+    {
+        parent::init();
+        $this->addServers($this->getMemcached(), $this->getServers());
+    }
+
+    /**
+     * Add servers to the server pool of the cache specified
+     *
+     * @param \Memcached $cache
+     * @param MemCachedServer[] $servers
+     * @throws InvalidConfigException
+     */
+    protected function addServers($cache, $servers)
+    {
+        if (empty($servers)) {
+            $servers = [new MemCachedServer([
+                'host' => '127.0.0.1',
+                'port' => 11211,
+            ])];
+        } else {
+            foreach ($servers as $server) {
+                if ($server->host === null) {
+                    throw new InvalidConfigException("The 'host' property must be specified for every memcached server.");
+                }
+            }
+        }
+
+        $existingServers = [];
+        if ($this->persistentId !== null) {
+            foreach ($cache->getServerList() as $s) {
+                $existingServers[$s['host'] . ':' . $s['port']] = true;
+            }
+        }
+        foreach ($servers as $server) {
+            if (empty($existingServers) || !isset($existingServers[$server->host . ':' . $server->port])) {
+                $cache->addServer($server->host, $server->port, $server->weight);
+            }
+        }
+    }
+
+    /**
+     * Returns the underlying memcached object.
+     * @return \Memcached the memcached object used by this cache component.
+     * @throws InvalidConfigException if memcached extension is not loaded
+     */
+    public function getMemcached()
+    {
+        if ($this->_cache === null) {
+            if (!extension_loaded('memcached')) {
+                throw new InvalidConfigException('MemCached requires PHP memcached extension to be loaded.');
+            }
+
+            $this->_cache = $this->persistentId !== null ? new \Memcached($this->persistentId) : new \Memcached;
+            if ($this->username !== null || $this->password !== null) {
+                $this->_cache->setOption(\Memcached::OPT_BINARY_PROTOCOL, true);
+                $this->_cache->setSaslAuthData($this->username, $this->password);
+            }
+            if (!empty($this->options)) {
+                $this->_cache->setOptions($this->options);
+            }
+        }
+
+        return $this->_cache;
+    }
+
+    /**
+     * Returns the memcached server configurations.
+     * @return MemCachedServer[] list of memcached server configurations.
+     */
+    public function getServers()
+    {
+        return $this->_servers;
+    }
+
+    /**
+     * @param array $config list of memcached server configurations. Each element must be an array
+     * with the following keys: host, port, persistent, weight, timeout, retryInterval, status.
+     * @see http://php.net/manual/en/memcached.addserver.php
+     */
+    public function setServers($config)
+    {
+        foreach ($config as $c) {
+            $this->_servers[] = new MemCachedServer($c);
+        }
+    }
+
+    /**
+     * Retrieves a value from cache with a specified key.
+     * This is the implementation of the method declared in the parent class.
+     * @param string $key a unique key identifying the cached value
+     * @return mixed|false the value stored in cache, false if the value is not in the cache or expired.
+     */
+    protected function getValue($key)
+    {
+        return $this->_cache->get($key);
+    }
+
+    /**
+     * Retrieves multiple values from cache with the specified keys.
+     * @param array $keys a list of keys identifying the cached values
+     * @return array a list of cached values indexed by the keys
+     */
+    protected function getValues($keys)
+    {
+        return $this->_cache->getMulti($keys);
+    }
+
+    /**
+     * Stores a value identified by a key in cache.
+     * This is the implementation of the method declared in the parent class.
+     *
+     * @param string $key the key identifying the value to be cached
+     * @param mixed $value the value to be cached.
+     * @see [Memcached::set()](http://php.net/manual/en/memcached.set.php)
+     * @param int $duration the number of seconds in which the cached value will expire. 0 means never expire.
+     * @return bool true if the value is successfully stored into cache, false otherwise
+     */
+    protected function setValue($key, $value, $duration)
+    {
+        // Use UNIX timestamp since it doesn't have any limitation
+        // @see http://php.net/manual/en/memcached.expiration.php
+        $expire = $duration > 0 ? $duration + time() : 0;
+
+        return $this->_cache->set($key, $value, $expire);
+    }
+
+    /**
+     * Stores multiple key-value pairs in cache.
+     * @param array $data array where key corresponds to cache key while value is the value stored
+     * @param int $duration the number of seconds in which the cached values will expire. 0 means never expire.
+     * @return array array of failed keys.
+     */
+    protected function setValues($data, $duration)
+    {
+        // Use UNIX timestamp since it doesn't have any limitation
+        // @see http://php.net/manual/en/memcached.expiration.php
+        $expire = $duration > 0 ? $duration + time() : 0;
+
+        // Memcached::setMulti() returns boolean
+        // @see http://php.net/manual/en/memcached.setmulti.php
+        return $this->_cache->setMulti($data, $expire) ? [] : array_keys($data);
+    }
+
+    /**
+     * Stores a value identified by a key into cache if the cache does not contain this key.
+     * This is the implementation of the method declared in the parent class.
+     *
+     * @param string $key the key identifying the value to be cached
+     * @param mixed $value the value to be cached
+     * @see [Memcached::set()](http://php.net/manual/en/memcached.set.php)
+     * @param int $duration the number of seconds in which the cached value will expire. 0 means never expire.
+     * @return bool true if the value is successfully stored into cache, false otherwise
+     */
+    protected function addValue($key, $value, $duration)
+    {
+        // Use UNIX timestamp since it doesn't have any limitation
+        // @see http://php.net/manual/en/memcached.expiration.php
+        $expire = $duration > 0 ? $duration + time() : 0;
+
+        return $this->_cache->add($key, $value, $expire);
+    }
+
+    /**
+     * Deletes a value with the specified key from cache
+     * This is the implementation of the method declared in the parent class.
+     * @param string $key the key of the value to be deleted
+     * @return bool if no error happens during deletion
+     */
+    protected function deleteValue($key)
+    {
+        return $this->_cache->delete($key, 0);
+    }
+
+    /**
+     * Deletes all values from cache.
+     * This is the implementation of the method declared in the parent class.
+     * @return bool whether the flush operation was successful.
+     */
+    protected function flushValues()
+    {
+        return $this->_cache->flush();
+    }
+}
