+++ framework/classes.php
@@ -72,8 +72,8 @@ return [
   'yii\caching\ExpressionDependency' => YII2_PATH . '/caching/ExpressionDependency.php',
   'yii\caching\FileCache' => YII2_PATH . '/caching/FileCache.php',
   'yii\caching\FileDependency' => YII2_PATH . '/caching/FileDependency.php',
-  'yii\caching\MemCache' => YII2_PATH . '/caching/MemCache.php',
-  'yii\caching\MemCacheServer' => YII2_PATH . '/caching/MemCacheServer.php',
+  'yii\caching\MemCached' => YII2_PATH . '/caching/MemCached.php',
+  'yii\caching\MemCachedServer' => YII2_PATH . '/caching/MemCachedServer.php',
   'yii\caching\TagDependency' => YII2_PATH . '/caching/TagDependency.php',
   'yii\caching\WinCache' => YII2_PATH . '/caching/WinCache.php',
   'yii\captcha\Captcha' => YII2_PATH . '/captcha/Captcha.php',
